Add files via upload
a couple of bugs and typos so the code is altered and some parts are left out for now. setting the location manually in the code. tensorflow 2.19.1. 

$ pip install pydot
then restart the jupyter lab. 

Download the titanic data manually from here:
https://www.openml.org/search?type=data&sort=runs&id=40945&status=active
convert to csv online:
https://pulipulichen.github.io/jieba-js/weka/arff2csv/
altered the columns on demand. preprocessing the data basically.